`tools/importer-rest-api-specs`: refactoring the Example Generation logic into the Terraform package
