+++ tools/importer-rest-api-specs/components/terraform/examples/resource.go
