+++ tools/importer-rest-api-specs/components/terraform/examples/variable_replacements.go
